# inline hash function to save extra alloc