--- a/internal/utils.go
+++ b/internal/utils.go
@@ -2,7 +2,6 @@ package internal
 
 import (
 	"fmt"
-	"hash/fnv"
 	"os"
 	"path/filepath"
 	"sort"
@@ -10,10 +9,18 @@ import (
 	"strings"
 )
 
+const (
+	offset64 = 14695981039346656037
+	prime64  = 1099511628211
+)
+
 func Hash(key []byte) uint64 {
-	h := fnv.New64a()
-	h.Write(key)
-	return h.Sum64()
+	var s uint64 = offset64
+	for _, c := range key {
+		s ^= uint64(c)
+		s *= prime64
+	}
+	return s
 }
 
 func Exists(path string) bool {